task_bot/requirements.txt
django==6.0
djangorestframework==3.16.1
redis==7.1.0
celery==5.6.0
psycopg==3.3.2
psycopg2-binary==2.9.11
asyncpg==0.31.0
gunicorn==23.0.0
aiogram==3.23.0
aiogram-dialog==2.4.0
aiohttp
python-dotenv==1.2.1
isort==7.0.0
flake8==7.3.0
